Move panels to h.panels package
This commit tidies up our use of "panels". While similar to views, these
are different things, so I've moved them to their own package.

In addition, we should probably only be using panels in situations where
we need to do additional computation beyond simply rendering a template,
so I've removed the "group_invite" panel and just inlined it back into
the one place it was being used.
