+++ h/panels/navbar.py
@@ -1,57 +1,12 @@
 # -*- coding: utf-8 -*-
 
-"""Shared components used across multiple pages on the site."""
+"""The navigation bar displayed at the top of most pages."""
 
 from __future__ import unicode_literals
 
 from pyramid_layout.panel import panel_config
 
-from h import i18n
-from h._compat import urlparse
-
-_ = i18n.TranslationString
-
-
-def _matches_route(path, request, route_name):
-    """
-    Return ``True`` if ``path`` matches the URL pattern for a given route.
-    """
-
-    introspector = request.registry.introspector
-
-    # `route` is a pyramid.interfaces.IRoute
-    route = introspector.get('routes', route_name)['object']
-    return route.match(path) is not None
-
-
-@panel_config(name='group_invite',
-              renderer='h:templates/panels/group_invite.html.jinja2')
-def group_invite(context, request, group_url):
-    return {'group_url': group_url}
-
-
-@panel_config(name='back_link',
-              renderer='h:templates/panels/back_link.html.jinja2')
-def back_link(context, request):
-    """
-    A link which takes the user back to the previous page on the site.
-    """
-
-    referrer_path = urlparse.urlparse(request.referrer or '').path
-    current_username = request.authenticated_user.username
-
-    if referrer_path == request.route_path('activity.user_search',
-                                           username=current_username):
-        back_label = _('Back to your profile page')
-    elif _matches_route(referrer_path, request, 'group_read'):
-        back_label = _('Back to group overview page')
-    else:
-        back_label = None
-
-    return {
-        'back_label': back_label,
-        'back_location': request.referrer,
-    }
+from h.i18n import TranslationString as _  # noqa
 
 
 @panel_config(name='navbar', renderer='h:templates/panels/navbar.html.jinja2')
@@ -71,7 +26,7 @@ def navbar(context, request, opts={}):
                 'link': request.route_url('group_read', pubid=group.pubid, slug=group.slug)
                 })
         user_activity_url = request.route_url('activity.user_search',
-            username=request.authenticated_user.username)
+                                              username=request.authenticated_user.username)
         username = request.authenticated_user.username
 
     if request.matched_route.name in ['group_read', 'activity.user_search']:
