--- h/templates/panels/group_invite.html.jinja2
@@ -1,13 +1,0 @@
-<h3 class="group-invite__title">
-  {% trans %}Invite new members{% endtrans %}
-</h3>
-Sharing the link below lets people join this group:
-<div class="group-invite__container js-copy-button">
-  <input class="group-invite__input js-select-onfocus"
-         data-ref="input"
-         value="{{ group_url }}">
-  <button class="group-invite__clipboard-button"
-          data-ref="button"
-          title="Copy to clipboard">
-  {{ svg_icon('copy_to_clipboard') }}
-</div>
